cumulative updates since 25.12 to align with cuML 26.06 (#1019)
In addition:
- dropping support for Spark 3.3 due to incompatible min python version
for cuML
- drops DB 13.3 and 14.3 support in benchmark and notebook examples

---------

+++ python/tests/test_metrics.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2024, NVIDIA CORPORATION.
+# Copyright (c) 2024-2026, NVIDIA CORPORATION.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -96,7 +96,7 @@ def test_multi_class_metrics(
     ).astype(np.float64)
 
     probabilities = np.random.rand(1000, num_classes)
-    probabilities[range(1000), list(pdf["label"].astype(np.integer))] = 2.0
+    probabilities[range(1000), list(pdf["label"].astype(int))] = 2.0
     probabilities = probabilities / np.sum(probabilities, axis=1).reshape(-1, 1)
 
     pdf["probabilities"] = list(probabilities)
